# docs/openapi.yaml
...
info:
  title: LiveReview API
  version: 1.0.0
openapi: 3.0.0
paths:
  /api/v1/activities:
    get:
      description: GetRecentActivities handles the API endpoint for fetching recent activities
      operationId: GetRecentActivities
      parameters:
        - in: query
          name: limit
          required: true
          schema:
            type: string
        - in: query
          name: offset
          required: true
          schema:
            type: string
      responses:
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      tags:
        - Activities
...
servers:
  - url: http://localhost:8888